Admin/requirements.txt
kivy==2.2.0
kivymd==1.1.1
pyyaml
pillow
multitasking
watchdog
opencv-python
numpy
face_recognition
cryptography
pandas
openpyxl